fix(api): fix transitive imports causing mock/testing mishaps
